Cementerios_Mapas

+++ AtoGobMx/Models/Gabetas.cs
@@ -12,8 +12,8 @@ namespace AtoGobMx.Models
         public string NombrePropietario { get; set; }
         public int NumeroEspasios { get; set; }
         public float MetrosCorrespondientes { get; set; }
-        public float Longitud { get; set; }
-        public float Latitud { get; set; }
+        public decimal Longitud { get; set; }
+        public decimal Latitud { get; set; }
         public int EspaciosDisponibles { get; set; }
         public int? CementerioId { get; set; }
         public bool Archivado { get; set; }
